refactor(reducer): extract Socket reduce-block (SocketRouter)
The sole `.socketMessage` owner is now its own reduce-block. `.socketMessage`
carries the request/response `reply` FD handle, so the exhaustive
`domain(of:)` partition maps it to the new `.socket` domain and only
`socketReducer` guards on it — preventing any second match from handing the
CLI a doubled JSON line or a premature EOF.

The whole nested `switch message` (agent lifecycle, pane / workspace / file
/ layout commands, request/response verbs) plus the socket-only helpers
(handleSocketWorkspaceCreate/Move, handleSocketGroupCreate, replyPaneCreated,
refreshSyncGroupForWorkspace, replySyncStatus) move verbatim into
AppReducer+Socket.swift. Every reply.send / reply.close sequence and the
agentStopped / notification / paneMoveToWorkspace effect orderings are
byte-identical. Shared helpers (resolvePaneTarget, paneSendText, tailLines,
the graft + handleWeb* handlers) stay reachable via self; the cross-file move
only relaxes replyPaneCreated / replySyncStatus / static resolveWorkspace from
private to internal so their existing core callers still see them.
